fix(test): renamed paths
-renamed paths to solve error
